Правки по замечаниям к ДЗ-4
